Consertei a leitura, user search e a organização
Além de corrigir o tamanho da hashtable dos usuários, consertei a pesquisa dos jogadores avaliados por usuários, agora a única coisa que falta é colocar o limite da pesquisa por posição
