+++ Trabalho_Final.layout
@@ -2,29 +2,34 @@
 <CodeBlocks_layout_file>
 	<FileVersion major="1" minor="0" />
 	<ActiveTarget name="Debug" />
-	<File name="main.cpp" open="1" top="1" tabpos="1" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
+	<File name="estruturas.hpp" open="1" top="0" tabpos="2" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
 		<Cursor>
-			<Cursor1 position="10913" topLine="135" />
+			<Cursor1 position="3327" topLine="25" />
 		</Cursor>
 		<Folding>
-			<Collapse line="17" />
+			<Collapse line="18" />
 			<Collapse line="27" />
-			<Collapse line="34" />
-			<Collapse line="110" />
-			<Collapse line="125" />
-			<Collapse line="129" />
-			<Collapse line="146" />
-			<Collapse line="151" />
-			<Collapse line="165" />
-			<Collapse line="171" />
-			<Collapse line="174" />
-			<Collapse line="206" />
+			<Collapse line="36" />
+			<Collapse line="43" />
+			<Collapse line="122" />
+			<Collapse line="137" />
+			<Collapse line="141" />
+			<Collapse line="159" />
+			<Collapse line="164" />
+			<Collapse line="172" />
+		</Folding>
+	</File>
+	<File name="main.cpp" open="1" top="1" tabpos="1" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
+		<Cursor>
+			<Cursor1 position="1201" topLine="63" />
+		</Cursor>
+		<Folding>
+			<Collapse line="111" />
+			<Collapse line="117" />
+			<Collapse line="145" />
+			<Collapse line="173" />
+			<Collapse line="175" />
 			<Collapse line="214" />
-			<Collapse line="243" />
-			<Collapse line="276" />
-			<Collapse line="282" />
-			<Collapse line="301" />
-			<Collapse line="380" />
 		</Folding>
 	</File>
 </CodeBlocks_layout_file>
