+++ miniminirating.csv
@@ -1,20 +1,21 @@
-user_id,sofifa_id,rating;;;;;
-118046,256417,1.5;;;;;
-97952,256417,2.5;;;;;
-24379,256417,4.0;;;;;
-12389,256417,2.0;;;;;
-135376,256417,1.5;;;;;
-14183,256417,2.5;;;;;
-100105,256417,3.5;;;;;
-21872,256417,3.5;;;;;
-117660,256417,2.0;;;;;
-108952,256417,5.0;;;;;
-118241,256417,1.5;;;;;
-12142,256417,2.5;;;;;
-121987,256417,2.0;;;;;
-26229,256417,4.5;;;;;
-88572,256417,3.5;;;;;
-11923,256417,3.5;;;;;
-50393,256417,3.0;;;;;
-43113,256417,2.0;;;;;
-13414,256417,2.0;;;;;
+user_id,sofifa_id,rating;;;;
+118046,256417,1.5;;;;
+97952,256417,2.5;;;;
+24379,256417,4.0;;;;
+12389,256417,2.0;;;;
+135376,256417,1.5;;;;
+14183,256417,2.5;;;;
+100105,256417,3.5;;;;
+21872,256417,3.5;;;;
+117660,256417,2.0;;;;
+108952,256417,5.0;;;;
+118241,256417,1.5;;;;
+12142,256417,2.5;;;;
+121987,256417,2.0;;;;
+26229,256417,4.5;;;;
+88572,256417,3.5;;;;
+11923,256417,3.5;;;;
+50393,256417,3.0;;;;
+43113,256417,2.0;;;;
+13414,256417,2.0;;;;
+4, 256417, 5.0;;;;
